[Warnings] Fixed "Possible Dangling Reference Warning"
In GCC-13, a new warning was added to detect dangling references to
temporary variables. This is a good addition as these temporaries are
deleted the line they are created; but passing a reference to them can
cause issues.

The NDMatrix class in VTR used const_cast to avoid duplicate code;
however, this required a cast to a reference which may be tripping up
the warning detection code in the compiler. I do not think this is an
actual issue.

In truth, the const_casts make the code very hard to read and
understand, when all the non-const version is doing is exactly what the
const version was doing. Since these functions are so small and simple,
I just duplicated the code. I only did this in places where the warning
was coming up.